Add wrap flex for playlist and add Rubik font
