+++ client/src/App.css
@@ -36,7 +36,8 @@
   font-size: 28px;
   line-height: 33px;
   text-align: center;
-
+  width: 200px;
+  height: 48px;
   /* Text */
 
   color: #F8F8F8;
