+++ client/src/pages/Followers.jsx
@@ -82,7 +82,7 @@ export default function Followers(){
           return <Follower key={i} user={follower} />
         })
 
-        const newFollowersInfo = newFollowerComponents.length > 0 ? <div className="center display-container"> {newFollowerComponents} </div> : 'There is no new follower. Keep rolling!';
+        const newFollowersInfo = newFollowerComponents.length > 0 ? <div className="center display-container" style={{flexWrap: 'wrap'}}> {newFollowerComponents} </div> : 'There is no new follower. Keep rolling!';
 
         unFollowers.forEach((unFollower, i) => {
           unFollowerComponents.push(
@@ -90,7 +90,7 @@ export default function Followers(){
           )
         })
 
-        const unFollowersInfo = unFollowerComponents.length > 0 ? <div className="center display-container"> {unFollowerComponents} </div> : 'There is no unfollower. You are so popular!';
+        const unFollowersInfo = unFollowerComponents.length > 0 ? <div className="center display-container" style={{flexWrap: 'wrap'}}> {unFollowerComponents} </div> : 'There is no unfollower. You are so popular!';
 
         return (
           <div>
