+++ client/src/pages/Homepage.css
@@ -2,7 +2,7 @@
     position: absolute;
     width: 454px;
     height: 57px;
-    left: 59px;
+    left: 89px;
     top: 343px;
 
     font-style: normal;
@@ -20,7 +20,7 @@
 
 .primary-button {
     position: absolute;
-    left: 80px;
+    left: 93px;
     top: 454px;
 }
 
@@ -28,7 +28,7 @@
     position: absolute;
     width: 449px;
     height: 28px;
-    left: 495px;
+    left: 580px;
     top: 750px;
 
     font-style: normal;
@@ -45,7 +45,7 @@
     position: absolute;
     width: 371px;
     height: 28px;
-    left: 534px;
+    left: 600px;
     top: 801px;
 
     font-style: normal;
@@ -61,10 +61,10 @@
 
 #entrance-images img {
     position: absolute;
-    width: 784px;
-    height: 523px;
-    left: 533px;
-    top: 135px;
+    width: 850px;
+    height: 600px;
+    left: 900px;
+    top: 100px;
     webkit-transition: opacity 1s ease-in-out;
     -moz-transition: opacity 1s ease-in-out;
     -o-transition: opacity 1s ease-in-out;
